Moved BGL Dijkstra solver into separate file
