+++ descartes_light/solvers/include/descartes_light/solvers/bgl/bgl_solver.h
@@ -97,20 +97,4 @@ public:
                         const std::vector<typename StateEvaluator<FloatType>::ConstPtr>& state_eval) override;
 };
 
-/**
- * @brief BGL solver implementation that constructs vertices and edges in the build function and uses Dijkstra's
- * algorithm with a default visitor to search the graph
- */
-template <typename FloatType>
-class BGLDijkstraSolverVE : public BGLSolverBaseVE<FloatType>
-{
-public:
-  using BGLSolverBaseVE<FloatType>::BGLSolverBaseVE;
-
-  SearchResult<FloatType> search() override;
-};
-
-using BGLDijkstraSolverVEF = BGLDijkstraSolverVE<float>;
-using BGLDijkstraSolverVED = BGLDijkstraSolverVE<double>;
-
 }  // namespace descartes_light
