+++ descartes_light/test/benchmarks/benchmarks.cpp
@@ -3,7 +3,7 @@
 #include <descartes_light/test/solver_factory.h>
 // Solvers
 #include <descartes_light/solvers/ladder_graph/ladder_graph_solver.h>
-#include <descartes_light/solvers/bgl/bgl_solver.h>
+#include <descartes_light/solvers/bgl/bgl_dijkstra_solver.h>
 
 DESCARTES_IGNORE_WARNINGS_PUSH
 #include <boost/core/demangle.hpp>
@@ -80,7 +80,7 @@ int main(int, char**)
   benchmark(SolverFactory<LadderGraphSolverD>());
   benchmark(SolverFactory<LadderGraphSolverF>());
 
-  // BGL ladder graph
+  // BGL full Dijkstra solver
   benchmark(SolverFactory<BGLDijkstraSolverVED>());
   benchmark(SolverFactory<BGLDijkstraSolverVEF>());
