+++ descartes_light/test/include/descartes_light/test/impl/solver_factory.hpp
@@ -2,24 +2,22 @@
 
 #include <descartes_light/test/solver_factory.h>
 #include <descartes_light/solvers/ladder_graph/ladder_graph_solver.h>
-#include <descartes_light/solvers/bgl/bgl_solver.h>
+#include <descartes_light/solvers/bgl/bgl_dijkstra_solver.h>
 
 namespace descartes_light
 {
-// Ladder graph solver Factory
-template <typename FloatT>
-struct SolverFactory<LadderGraphSolver<FloatT>>
+// Ladder graph solver factory
+template <typename FloatType>
+struct SolverFactory<LadderGraphSolver<FloatType>>
 {
-  using FloatType = FloatT;
-  typename Solver<FloatT>::Ptr create() const { return std::make_unique<LadderGraphSolver<FloatT>>(1); }
+  typename Solver<FloatType>::Ptr create() const { return std::make_shared<LadderGraphSolver<FloatType>>(1); }
 };
 
-// Boost Ladder graph solver Factory
-template <typename FloatT>
-struct SolverFactory<BGLDijkstraSolverVE<FloatT>>
+// Boost full Dijkstra graph solver factory
+template <typename FloatType>
+struct SolverFactory<BGLDijkstraSolverVE<FloatType>>
 {
-  using FloatType = FloatT;
-  typename Solver<FloatT>::Ptr create() const { return std::make_unique<BGLDijkstraSolverVE<FloatT>>(1); }
+  typename Solver<FloatType>::Ptr create() const { return std::make_shared<BGLDijkstraSolverVE<FloatType>>(1); }
 };
 
 }  // namespace descartes_light
